feature: On meg stop, give option to stop all
Fixes megalus-1
