+++ megalus/bash/commands.py
@@ -25,7 +25,7 @@ import click
 from buzio import console
 from loguru import logger
 
-from megalus import Megalus
+from megalus.main import Megalus
 from megalus.utils import client, find_containers
