+++ megalus/cmd.py
@@ -1,11 +1,13 @@
 import click
 
-from megalus import Megalus
+from megalus.check.commands import check
+from megalus.main import Megalus
 from megalus.bash.commands import bash
 from megalus.build.commands import build, buildgroup
 from megalus.clone.commands import clone
 from megalus.commands.commands import config, install, reset, update
-from megalus.compose.commands import restart, scale, stop, up
+from megalus.compose.commands import restart, scale, up
+from megalus.stop.commands import stop
 from megalus.down.commands import down
 from megalus.logs.commands import logs
 from megalus.run.commands import run
@@ -35,6 +37,7 @@ cli.add_command(scale)
 cli.add_command(stop)
 cli.add_command(up)
 cli.add_command(update)
+cli.add_command(check)
 
 
 def start():
