+++ megalus/down/commands.py
@@ -1,7 +1,7 @@
 import click
 from loguru import logger
 
-from megalus import Megalus
+from megalus.main import Megalus
 
 
 @click.command()
