+++ megalus/run/commands.py
@@ -7,7 +7,7 @@ import click
 from buzio import console
 from loguru import logger
 
-from megalus import Megalus
+from megalus.main import Megalus
 from megalus.utils import run_command
